bot/requirements.txt
python-telegram-bot==20.7   #library of bot
SQLAlchemy==2.0.23          # ORM
alembic==1.13.1             #database migrations
python-dotenv==1.0.1        # read enviroment variables (.env)
psycopg2-binary==2.9.9
